Add support for log of updates in real-time mode
